# JIT policy eval and add T-Maze