--- a/grl/environment/examples_lib.py
+++ b/grl/environment/examples_lib.py
@@ -1,6 +1,7 @@
 import numpy as np
 
 from .memory_lib import *
+from .tmaze_lib import tmaze
 """
 Library of POMDP specifications. Each function returns a dict of the form:
     {
@@ -625,6 +626,15 @@ def simple_chain(n: int = 10):
     Pi_phi = [np.ones((n, 1))]
 
     return to_dict(T, R, 0.9, p0, phi, Pi_phi)
+
+
+def tmaze_5_two_thirds_up():
+    # n_obs x n_actions
+    n = 5
+    discount = 0.9
+    Pi_phi = [np.array([[0, 1, 0, 0], [0, 1, 0, 0], [0, 1, 0, 0], [2 / 3, 0, 1 / 3, 0]])]
+    return to_dict(*tmaze(n, discount=discount), Pi_phi)
+
 
 def to_dict(T, R, gamma, p0, phi, Pi_phi, Pi_phi_x=None):
     return {