--- a/grl/environment/examples_lib.py
+++ b/grl/environment/examples_lib.py
@@ -2,6 +2,7 @@ import numpy as np
 
 from .memory_lib import *
 from .tmaze_lib import tmaze, slippery_tmaze
+from grl.mdp import random_stochastic_matrix
 """
 Library of POMDP specifications. Each function returns a dict of the form:
     {
@@ -11,7 +12,7 @@ Library of POMDP specifications. Each function returns a dict of the form:
         p0:       starting state probabilities,
         phi:      observation matrix (currently the same for all actions),
         Pi_phi:   policies to evaluate in the observation space of the POMDP
-        T_mem:    memory transition function
+        mem_params:    memory parameters. We get our memory function by calling softmax(mem_params, axis=-1)
         Pi_phi_x: policies to evaluate in the observation space of the cross product of the underyling MDP and memory function
     }
 
@@ -693,6 +694,19 @@ def slippery_tmaze_5_two_thirds_up():
     discount = 0.9
     Pi_phi = [
         np.array([[0, 0, 1, 0], [0, 0, 1, 0], [0, 0, 1, 0], [2 / 3, 1 / 3, 0, 0], [1, 0, 0, 0]])
+    ]
+
+    # memory policy is observations * memory bits (2) x n_actions
+    Pi_phi_x = [Pi_phi[0].repeat(2, axis=0)]
+    return to_dict(*slippery_tmaze(n, discount=discount, slip_prob=0.3), Pi_phi, Pi_phi_x)
+
+def slippery_tmaze_5_random():
+    # n_obs x n_actions
+    n = 5
+    discount = 0.9
+    Pi_phi = [
+        random_stochastic_matrix((5, 4))
+        # np.array([[0, 0, 1, 0], [0, 0, 1, 0], [0, 0, 1, 0], [2 / 3, 1 / 3, 0, 0], [1, 0, 0, 0]])
     ]
 
     # memory policy is observations * memory bits (2) x n_actions