--- a/grl/environment/memory_lib.py
+++ b/grl/environment/memory_lib.py
@@ -1,3 +1,4 @@
+import sys
 import numpy as np
 """
 1 bit memory functions with three obs: r, b, t
@@ -6,7 +7,21 @@ and 2 actions: up, down
 Dimensions: AxZxMxM
 """
 
-mem_0 = np.array([
+def get_memory(memory_id: int, n_obs: int, n_actions: int, n_mem_states: int = 2) -> np.ndarray:
+    current_module = globals()
+    mem_name = f'memory_{memory_id}'
+    if memory_id == 0:
+        mem_params = np.random.normal(size=(n_actions, n_obs, n_mem_states, n_mem_states)) * np.sqrt(2)
+    else:
+        if mem_name in current_module:
+            T_mem = current_module[mem_name]
+            # smooth out for softmax
+            mem_params = np.log(T_mem + 1e-5)
+        else:
+            raise NotImplementedError(f'{mem_name} not found in memory_lib.py') from None
+    return mem_params
+
+mem_1 = np.array([
     [ # red
         # Pr(m'| m, o)
         # m0', m1'
@@ -22,7 +37,7 @@ mem_0 = np.array([
         [1, 0],
     ],
 ])
-memory_0 = np.array([mem_0, mem_0]) # up, down
+memory_1 = np.array([mem_1, mem_1]) # up, down
 
 mem_3 = np.array([
     [ # red
@@ -272,5 +287,4 @@ mem_17 = np.array([
 ])
 memory_17 = np.array([mem_17, mem_17, mem_17, mem_17]) # up, down, right, left
 
-mem_18 = np.random.normal(size=(5, 2, 2)) * np.sqrt(2)
-memory_18 = np.array([mem_18, mem_18, mem_18, mem_18]) # up, down, right, left
+