--- a/grl/environment/memory_lib.py
+++ b/grl/environment/memory_lib.py
@@ -11,7 +11,8 @@ def get_memory(memory_id: int, n_obs: int, n_actions: int, n_mem_states: int = 2
     current_module = globals()
     mem_name = f'memory_{memory_id}'
     if memory_id == 0:
-        mem_params = np.random.normal(size=(n_actions, n_obs, n_mem_states, n_mem_states)) * np.sqrt(2)
+        mem_params = np.random.normal(size=(n_actions, n_obs, n_mem_states,
+                                            n_mem_states)) * np.sqrt(2)
     else:
         if mem_name in current_module:
             T_mem = current_module[mem_name]
@@ -286,5 +287,3 @@ mem_17 = np.array([
     ],
 ])
 memory_17 = np.array([mem_17, mem_17, mem_17, mem_17]) # up, down, right, left
-
-