BIG PR ALERT!

We do a number of things in this PR. In descending order in terms of the size of the changes:

- `jit`ify the policy eval class. Things are super fast now! Especially for the `do_grad` function which needs to call `PolicyEval.run` a bunch. The biggest change is batchifying the functions in the class.
- Tests for policy eval and also for the analytical solution for value function learning. This is mainly done through the new `SimpleChain` environment, which is a 1-action 10-state MDP that is a chain that goes right to the next state, until it terminates at the last node with a reward of +1.
- Fixes for T-Maze. Most notably @dsctt mentioned that there needs to be a special observation for the terminal state - this is indeed very much true, and has been rectified for the T-Maze environment.
- Minor tweaks for the `MDP` class, to make things fast as well.

NOTE: I've been changing things so that `numpy` is `np` and `jax.numpy` is `jnp`.

<s>So all-in-all, our tests went from wall-clock time 83 seconds to 4 seconds on my local machine 🙂 </s>
edit: I'm an idiot and commented out the two grad tests
edit 2: tests went from 83s to 6s. I made sure to check all the tests are passing. Now analytical `do_grad` is much faster!